+++ llvm-hs/src/LLVM/Internal/FFI/OrcJIT.hs
@@ -10,8 +10,12 @@ import Foreign.Ptr
 import LLVM.Internal.FFI.DataLayout
 import LLVM.Internal.FFI.LLVMCTypes
 
+-- | Abstract type used as the identifier for a module.
+newtype ModuleKey = ModuleKey Word64 deriving (Eq, Ord, Show)
+
 data JITSymbol
-data LambdaResolver
+data SymbolResolver
+data ExecutionSession
 
 newtype TargetAddress = TargetAddress Word64
 
@@ -24,12 +28,12 @@ foreign import ccall safe "LLVM_Hs_disposeJITSymbol" disposeSymbol ::
   Ptr JITSymbol -> IO ()
 
 foreign import ccall safe "LLVM_Hs_createLambdaResolver" createLambdaResolver ::
+  Ptr ExecutionSession ->
   FunPtr SymbolResolverFn ->
-  FunPtr SymbolResolverFn ->
-  IO (Ptr LambdaResolver)
+  IO (Ptr SymbolResolver)
 
-foreign import ccall safe "LLVM_Hs_disposeLambdaResolver" disposeLambdaResolver ::
-  Ptr LambdaResolver -> IO ()
+foreign import ccall safe "LLVM_Hs_disposeSymbolResolver" disposeSymbolResolver ::
+  Ptr SymbolResolver -> IO ()
 
 foreign import ccall safe "LLVM_Hs_JITSymbol_getAddress" getAddress ::
   Ptr JITSymbol -> Ptr (OwnerTransfered CString) -> IO TargetAddress
@@ -48,3 +52,21 @@ foreign import ccall safe "LLVM_Hs_getMangledSymbol" getMangledSymbol ::
 
 foreign import ccall safe "LLVM_Hs_disposeMangledSymbol" disposeMangledSymbol ::
   CString -> IO ()
+
+foreign import ccall safe "LLVM_Hs_createExecutionSession" createExecutionSession ::
+  IO (Ptr ExecutionSession)
+
+foreign import ccall safe "LLVM_Hs_disposeExecutionSession" disposeExecutionSession ::
+  Ptr ExecutionSession -> IO ()
+
+foreign import ccall safe "LLVM_Hs_allocateVModule" allocateVModule ::
+  Ptr ExecutionSession -> IO ModuleKey
+
+foreign import ccall safe "LLVM_Hs_releaseVModule" releaseVModule ::
+  Ptr ExecutionSession -> ModuleKey -> IO ()
+
+foreign import ccall "wrapper" wrapGetSymbolResolver ::
+  (ModuleKey -> IO (Ptr SymbolResolver)) -> IO (FunPtr (ModuleKey -> IO (Ptr SymbolResolver)))
+
+foreign import ccall "wrapper" wrapSetSymbolResolver ::
+  (ModuleKey -> Ptr SymbolResolver -> IO ()) -> IO (FunPtr (ModuleKey -> Ptr SymbolResolver -> IO ()))